Replace the modifying-index.html hackery with a view for both cases
This loses the ability to view the site without a web server running, as the front page is now served by inserting content into a view instead of as a static page. This was more important when I was just doing the design, but that's pretty much over with now.

(Besides, without a web server running, you're not going to be able to get any results)